+++ backend/src/main/java/com/mpl/backend/repository/PruebaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mpl.backend.model.Prueba;
 
-public interface PruebaRepository extends JpaRepository<Prueba, Integer> {
+public interface PruebaRepository extends JpaRepository<Prueba, Long> {
 }
